Change floats to doubles, implemented preconditioned CG

+++ src/comp559/fluid/LinearSolver.java
@@ -1,6 +1,6 @@
 package comp559.fluid;
 
-import org.jblas.FloatMatrix;
+import org.jblas.DoubleMatrix;
 
 /**
  * @author litalien
@@ -19,21 +19,11 @@ public abstract class LinearSolver {
     /** Max error for algorithm */
     public double maxError;
 
-    // public int IX( int i, int j, int N ) { return i*(N+2) + j; }
-
     /** Default constructor */
     public LinearSolver() {
         this.iteration = 0;
         this.resNorm2 = Double.NaN;
     }
 
-//    public void updatePressures( FloatMatrix x, float[] p, int N ) {
-//        for (int i = 1; i <= N; i++) {
-//            for (int j = 1; j <= N; j++) {
-//                p[IX(i, j, N)] = x.get(i-1 + N*(j-1));
-//            }
-//        }
-//    }
-
-    public abstract void solve( FloatMatrix A, FloatMatrix b, FloatMatrix x );
+    public abstract void solve( DoubleMatrix A, DoubleMatrix b, DoubleMatrix x );
 }
